+++ src/App.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import {
   BicyclingLayer,
+  GroundOverlay,
   MapBox,
   OverlayView,
   TrafficLayer,
@@ -48,6 +49,17 @@ const App = () => {
         <BicyclingLayer />
         <TrafficLayer />
         <TransitLayer />
+        <GroundOverlay
+          opts={{
+            url: 'https://placehold.it/256x256',
+            bounds: {
+              east: -73.98,
+              west: -73.985,
+              north: 40.706,
+              south: 40.702,
+            },
+          }}
+        />
       </div>
     </div>
   )